assignments/leaderboard/requirements.txt
numpy
pandas
pygithub
scikit-learn
tqdm